requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
dukpy==0.5.0
et_xmlfile==2.0.0
idna==3.10
jdcal==1.4.1
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
setuptools==75.8.0
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
uv==0.7.2
you-get==0.4.1730
